Use iOS 11 UISearchController to make it compatible with iPhone X
There were an issues with search bar on iPhone X, and the easiest way to fix it
was to use the newest UISearchController. It looks nice and behaves well, plus
it allowed to remove some offset-jugglery with header view.

It also needs to use iOS 11 as a minimum, but I don't think that's a problem
with this kind of app.